Allow "dumb" completion in the shell.
For when I'm doing dumb things, I need dumb completion to help my tiny
brain do them anyways. Inspired by needing to add -Nf files to a git
repo (this git repo) that are ignored, and weeping madly when the
instant-gratification of the bash completion that I've become so used to
(dependent on?) deserted me. Say it again with me: pampered, pampered
young man.