[mail] Configure offlineimap for awesomeness.
Another exceptional tool, and the keyring.py is a gift to once more
avoid the plaintexting of my passwords. Which I'm fond of.

The configuration itself is pretty basic, and currently it does not work
with the USC webmail account, because of a bizarre issue with actually
getting the key into the keyring somehow. Anyways, everything else is
pretty much straight sync, modulo excluding some of the pain-in-my-ass
gmail folders. Consider adding filters/namemapping for some of the lists
to avoid the huge namespaces, but will do later.