Segregate repo from junk in $HOME.
Trying to manage everything directly in the root was already starting to
scale poorly, and was a recipe for both trouble and forgotten pieces.
Now the dotfiles repo is normal, and I've symlinked the config files to
the managed ones (which should stay updated fine through pulls here).
The added benefit here is that instead of hidden files, they are now
normal ones.