[mail] Ignore all offlineimap and mutt caches.
Letting offlineimap do what it does best and sync my mail is very
desired in all locations, and not at all something I want to do via git.
Ditto for header caches; with both of these it should be allowed that
each machine is in a different state, even if the configs are updated.