[mail] Switch 'check-mail' macro to use offlineimap.
I don't use fetchmail any more for any of the accounts I'm currently
concerned about, and don't want to risk inadvertently overwriting stuff
with any of its bullshit when it's so nicely managed by offlineimap.

This also fixes the annoyance of having to open a separate terminal to
run offlineimap constantly (particularly given its fragility and the
fragility of my WLAN in the apartment in Bonn).