Add some config tweaks for gVim.
Fixing the font resolution on my laptop leads to wanting a slightly
smaller default font. Also, I'm a colorscheme ADD kid.

The cursorline goes here and not in the general configuration because,
sadly, in the 8-bit terminal land, it is a hideous underline that is
more distracting than helpful. It looks lovely with monokai though...

guioptions+=c means that rather than getting the hideous GTK1 popups I
so loathe, we use console dialogs, the way vim was meant to be
interacted with.